+++ dom/media/webaudio/PannerNode.h
@@ -1,5 +1,3 @@
-/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
-/* vim:set ts=2 sw=2 sts=2 et cindent: */
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -64,8 +62,8 @@ class PannerNode final : public AudioNode {
     SendInt32ParameterToTrack(DISTANCE_MODEL, int32_t(mDistanceModel));
   }
 
-  void SetPosition(double aX, double aY, double aZ, ErrorResult& aRv);
-  void SetOrientation(double aX, double aY, double aZ, ErrorResult& aRv);
+  void SetPosition(float aX, float aY, float aZ, ErrorResult& aRv);
+  void SetOrientation(float aX, float aY, float aZ, ErrorResult& aRv);
 
   double RefDistance() const { return mRefDistance; }
   void SetRefDistance(double aRefDistance, ErrorResult& aRv) {
