+++ dom/storage/SessionStorageManager.cpp
@@ -1,5 +1,3 @@
-/* -*- Mode: C++; tab-width: 8; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
-/* vim: set ts=8 sts=2 et sw=2 tw=80: */
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -655,8 +653,7 @@ void SessionStorageManager::ClearStorages(
       return;
     }
 
-    mActor->SendClearStorages(aPattern, nsCString(aOriginScope),
-                              static_cast<uint32_t>(aMode));
+    mActor->SendClearStorages(aPattern, nsCString(aOriginScope), aMode);
   }
 
   ClearStoragesInternal(aPattern, aOriginScope, aMode);
